requirements.txt
alembic>=1.13,<2.0
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
colorama==0.4.6
fastapi==0.110.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0

# PostgreSQL — ТОЛЬКО psycopg v3
psycopg[binary]>=3.1

pydantic==2.6.1
pydantic-settings==2.2.1
pydantic_core==2.16.2
pytest==8.0.0
python-dotenv==1.0.1
python-jose[cryptography]
passlib[bcrypt]

SQLAlchemy==2.0.25
starlette==0.36.3
typing_extensions==4.15.0
uvicorn==0.27.1
